requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
croniter==2.0.5
discord-py-interactions==5.13.1
discord-typings==0.9.0
discord.py==2.4.0
emoji==2.12.1
Faker==27.0.0
fore==0.1.8
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyotp==2.9.0
PySimpleGUI==5.0.6
PySocks==1.7.1
pystyle==2.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tk==0.1.0
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
PyNaCl==1.5.0
audioop-lts==0.1.0